requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
click==8.1.8
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
docopt==0.6.2
Flask==3.1.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-WTF==1.2.2
idna==3.10
intervaltree==3.1.0
ipymarkup==0.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
MarkupSafe==3.0.2
marshmallow==3.26.1
msgspec==0.19.0
natasha==1.6.0
navec==0.10.0
numpy==2.2.4
packaging==24.2
py-sc-client==0.4.0
py-sc-kpm==0.4.0
pydantic==2.11.3
pydantic_core==2.33.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
razdel==0.5.0
redis==5.2.1
setuptools==80.1.0
slovnet==0.6.0
sortedcontainers==2.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
websocket-client==1.8.0
Werkzeug==3.1.3
WTForms==3.2.1
yargy==0.16.0
